pyproject.toml
[project]
name = "quarry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "spotipy>=2.25.2",
]
